Small edits, adding counts, changing Boat length to float

+++ utils/parser_functions.py
@@ -1,11 +1,11 @@
 import re
 
 
-def parse_length(length_text: str) -> int:
+def parse_length(length_text: str) -> float:
     match = re.search(r'(\d+)\s*ft', length_text, re.IGNORECASE)
     if not match:
         raise ValueError(f"Could not parse boat length from: '{length_text}'")
-    return int(match.group(1))
+    return float(match.group(1))
 
 def parse_price(price_text: str) -> float:
     cleaned = price_text.lower()
@@ -18,4 +18,3 @@ def parse_price(price_text: str) -> float:
         raise ValueError(f"Could not parse price from: '{price_text}'")
 
     return float(match.group(1))
-
